Refactor filing persistence tests to use cluster sharding
